+++ packages/create-hydrogen-app/CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+- Devcontainer support added [#164](https://github.com/Shopify/hydrogen/pull/164)
+- fix: add check for products.length in `Welcome.server.jsx`
 
 ## 0.6.4 - 2021-11-11