+++ packages/hydrogen/CHANGELOG.md
@@ -7,6 +7,9 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+- fix: default to `retry: false` in `useQuery`
+- fix: starter template media gallery error when handling videos
+- fix: run graphiql middleware before vite, fixing graphiql
 - feat: allow custom Model3D poster
 
 ## 0.6.4 - 2021-11-11